crowdworking: moved fallback into own package
